\name{ausgdp}
\alias{ausgdp}
\docType{data}
\title{Quarterly Australian GDP}
\description{Quarterly Australian GDP per capita, 1971:1 - 1998:1}
\usage{data(ausgdp)}
\format{time series}
\source{
  Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) \emph{Forecasting with exponential smoothing: the state space approach},
  Springer. 
}
\references{
 \url{http://www.exponentialsmoothing.net}
}
\examples{
plot(ausgdp,main="Australian GDP per capita",ylab="dollars",xlab="Year")
}
\keyword{datasets}

