% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{carparts}
\alias{carparts}
\title{Monthly sales car parts}
\format{
multiple time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\usage{
carparts
}
\description{
Monthly sales car parts.  2674 series. Jan 1998 - Mar 2002.
}
\examples{

plot(carparts[, 2001:2010], main = "Monthly car part sales", xlab = "Year")

}
\references{
\url{https://robjhyndman.com/expsmooth/}
}
\keyword{datasets}
