% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{djiclose}
\alias{djiclose}
\title{Monthly Dow Jones Index: closing}
\format{
multiple time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
djiclose
}
\description{
Closing values of the Dow Jones Index on the first day of each month,
October 1928 - Dec 2007.  Two columns: \code{close} and \code{pcreturn}.
}
\examples{

plot(djiclose,main="Dow Jones Index",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
