% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{package}
\name{expsmooth-package}
\alias{expsmooth-package}
\alias{expsmooth}
\title{Data sets for "Forecasting with exponential smoothing"}
\description{
Data sets from the book "Forecasting with exponential smoothing: the state
space approach" by Hyndman, Koehler, Ord and Snyder (Springer, 2008).
}
\references{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D.,
(2008) \emph{Forecasting with exponential smoothing: the state space
approach}, Springer. \url{https://robjhyndman.com/expsmooth/}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://pkg.robjhyndman.com/expsmooth/}
  \item \url{https://github.com/robjhyndman/expsmooth}
  \item Report bugs at \url{https://github.com/robjhyndman/expsmooth/issues}
}

}
\author{
Rob J Hyndman.
}
\keyword{internal}
