% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{fmsales}
\alias{fmsales}
\title{Weekly FM sales}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
fmsales
}
\description{
Sales of a product for 62 weeks starting in early 2003.
}
\examples{

plot(fmsales,ylab="FM sales (thousands)",xlab="Week")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
