% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{freight}
\alias{freight}
\title{Annual US new freight cars}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer. \url{www.exponentialsmoothing.net}
}
\description{
Annual US new freight cars, 1947-1993. Freight cars, new (excl. rebuilt),
new orders, equip. mfrers. Series N0193 from the M3 competition.
}
\examples{

plot(freight,main="New freight cars shipped in USA",xlab="Year",ylab="")

}
\keyword{datasets}
