% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{hospital}
\alias{hospital}
\title{Monthly patient count}
\format{multiple time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\description{
Monthly patient count for products that are related to medical problems.
There are 767 time series that had a mean count of at least 10 and no zeros.
}
\examples{

plot(hospital[,1:10],main="Monthly patient count",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
