% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{mcopper}
\alias{mcopper}
\title{Monthly copper prices}
\format{
time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\usage{
mcopper
}
\description{
Monthly copper prices. Copper, grade A, electrolytic wire
bars/cathodes,LME,cash (pounds/ton) Source: UNCTAD
(\url{http://stats.unctad.org/Handbook}).
}
\examples{

plot(mcopper,main="Monthly copper price",ylab="pounds per ton",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
