% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{msales}
\alias{msales}
\title{Monthly product sales}
\format{
time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\usage{
msales
}
\description{
Monthly sales for a product with shortage indicators. Contains sales (first
column) and stockout indicator (second column).
}
\examples{

plot(msales[, 1], main = "Monthly sales of a product", ylab = "Sales", xlab = "Year")
points(msales, pch = (msales[, "stockouts"] == 1) + 1)
legend("bottomright", pch = 1:2, legend = c("Excess stock", "Stock shortage"))

}
\references{
\url{https://robjhyndman.com/expsmooth/}
}
\keyword{datasets}
