\name{partx}
\alias{partx}
\docType{data}
\title{Monthly sales of an automobile part}
\description{
  Monthly sales of an automobile part.
}
\usage{data(partx)}
\format{time series}
\source{
  Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) \emph{Forecasting with exponential smoothing: the state space approach},
  Springer
}
\references{
 \url{http://www.exponentialsmoothing.net}
}
\examples{
plot(partx,main="Monthly sales of an automobile part",ylab="Sales",xlab="Year")
}
\keyword{datasets}


