\name{usgdp}
\alias{usgdp}
\docType{data}
\title{ Quarterly US GDP}
\description{
  Quarterly US GDP. 1947:1 - 2006.1.
}
\usage{data(usgdp)}
\format{time series}
\source{
  Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) \emph{Forecasting with exponential smoothing: the state space approach},
  Springer
}
\references{
 \url{http://www.exponentialsmoothing.net}
}
\examples{
plot(usgdp,main="Quarterly US GDP",xlab="Year",ylab="US Dollars")
}
\keyword{datasets}


