% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{usnetelec}
\alias{usnetelec}
\title{Annual US net electricity generation}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\description{
Annual US net electricity generation (billion kwh) for 1949-2003
}
\examples{

plot(usnetelec,main="Annual US net electricity generation",ylab="billion kwh",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
