\name{visitors}
\alias{visitors}
\docType{data}
\title{Monthly Australian overseas vistors}
\description{Monthly Australian short-term overseas vistors. May 1985-April 2005}

\usage{data(visitors)}
\format{time series}
\source{
  Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) \emph{Forecasting with exponential smoothing: the state space approach},
  Springer.
}
\references{
 \url{http://www.exponentialsmoothing.net}
}
\examples{
plot(visitors,main="Overseas visitors to Australia",ylab="Thousands of people",xlab="Year")
}
\keyword{datasets}

