% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{ausgdp}
\alias{ausgdp}
\title{Quarterly Australian GDP}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
ausgdp
}
\description{
Quarterly Australian GDP per capita, 1971:1 - 1998:1
}
\examples{

plot(ausgdp,main="Australian GDP per capita",ylab="dollars",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
