% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{bonds}
\alias{bonds}
\title{Monthly US government bond yields}
\format{
time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
bonds
}
\description{
Monthly US government 10-year bond yields (percent pa) from Jan 1994 to May
2004
}
\examples{

plot(bonds, main = "US 10-year bonds yield", ylab = "Percentage per annum", xlab = "Year")

}
\references{
\url{https://robjhyndman.com/expsmooth/}
}
\keyword{datasets}
