% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{cangas}
\alias{cangas}
\title{Monthly Canadian gas production}
\format{
time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
cangas
}
\description{
Monthly Canadian gas production, billions of cubic metres, January 1960 -
February 2005
}
\examples{

plot(cangas, main = "Monthly Canadian gas production", ylab = "billion cubic metres", xlab = "Year")

}
\references{
\url{https://robjhyndman.com/expsmooth/}
}
\keyword{datasets}
