% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{enplanements}
\alias{enplanements}
\title{Monthly US domestic enplanements}
\format{
time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\usage{
enplanements
}
\description{
"Domestic Revenue Enplanements (millions): 1996-2000.  SOURCE: Department of
Transportation, Bureau of Transportation Statistics, Air Carrier Traffic
Statistic Monthly.
}
\examples{

plot(enplanements, main = "US domestic enplanements", ylab = "millions", xlab = "Year")

}
\references{
\url{https://robjhyndman.com/expsmooth/}
}
\keyword{datasets}
