\name{expsmooth-package}
\alias{expsmooth-package}
\alias{expsmooth}
\docType{package}
\title{
Data sets for "Forecasting with exponential smoothing"
}
\description{Data sets from the book "Forecasting with exponential smoothing: the state space approach" by 
	Hyndman, Koehler, Ord and Snyder (Springer, 2008).
}

\author{Rob J Hyndman. \email{Rob.Hyndman@buseco.monash.edu.au}}

\references{Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008) \emph{Forecasting with exponential smoothing: the state space approach},
  Springer. \url{www.exponentialsmoothing.net}.
}

\keyword{ package }


