% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{jewelry}
\alias{jewelry}
\title{Weekly jewelry sales}
\format{multiple time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
jewelry
}
\description{
Weekly sales of 314 costume jewelry items over the period week 5, 1998 to
week 24, 2000.
}
\examples{

plot(jewelry[,1:10],main="Weekly sales of costume jewelry items",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
