% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{partx}
\alias{partx}
\title{Monthly sales of an automobile part}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\usage{
partx
}
\description{
Monthly sales of an automobile part.
}
\examples{

plot(partx,main="Monthly sales of an automobile part",ylab="Sales",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
