% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{ukcars}
\alias{ukcars}
\title{Quarterly UK passenger car production}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\usage{
ukcars
}
\description{
Quarterly UK passenger car production (thousands of cars). 1977:1-2005:1
}
\examples{

plot(ukcars,main="UK passenger vehicle production",ylab="Thousands of cars",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
