% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{usgdp}
\alias{usgdp}
\title{Quarterly US GDP}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer
}
\description{
Quarterly US GDP. 1947:1 - 2006.1.
}
\examples{

plot(usgdp,main="Quarterly US GDP",xlab="Year",ylab="US Dollars")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
