% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{utility}
\alias{utility}
\title{Hourly utility demand}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\description{
Hourly utility demand, mid western USA from 1 Jan 2003
}
\examples{

plot(utility,main="Hourly utility demand",ylab="MW",xlab="Day")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
