% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{vehicles}
\alias{vehicles}
\title{Hourly vehicle counts}
\format{
time series
}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\usage{
vehicles
}
\description{
Hourly vehicle counts on Monash Freeway, outside Melbourne in Victoria,
Australia, beginning August 1995.
}
\examples{

plot(vehicles, main = "Hourly vehicle count", xlab = "Day", ylab = "Number of vehicles")

}
\references{
\url{https://robjhyndman.com/expsmooth/}
}
\keyword{datasets}
