% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{visitors}
\alias{visitors}
\title{Monthly Australian overseas vistors}
\format{time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\description{
Monthly Australian short-term overseas vistors. May 1985-April 2005
}
\examples{

plot(visitors,main="Overseas visitors to Australia",ylab="Thousands of people",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
