% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/expsmooth-package.R
\docType{data}
\name{xrates}
\alias{xrates}
\title{Monthly exchange rates}
\format{multiple time series}
\source{
Hyndman, R.J., Koehler, A.B., Ord, J.K., and Snyder, R.D., (2008)
\emph{Forecasting with exponential smoothing: the state space approach},
Springer.
}
\description{
Monthly foreign exchange rates: US dollar and UK pound against the Australia
dollar. \code{audusd} contains USD/AUD and \code{audukp} contains UKP/AUD.
}
\examples{

plot(xrates,main="Foreign exchange rates",xlab="Year")

}
\references{
\url{http://www.exponentialsmoothing.net}
}
\keyword{datasets}
